refactor: move auth to config
